Invalid transitions should raise NotImplementedError.
`automat._core.Automaton.outputForInput` implies that an invalid
transition should raise a `NotImplementedError`.  In practice, though,
`MethodicalMachine` may raise a `KeyError` because there's no
collector specified for a given state, because no transitions exist.

Reorder the code to always raise a `NotImplementedError`, add a test
to ensure this, and change the repr methods of various machine
components to succinctly describe themselves.